Update animation logic and enhance display in tape-transation-infinite
This commit includes updating the animation logic in infiniteTapeCore.js to cater for different configurations and conditions, as well as enhancing the display of the infiniteTape demo.html. The animation timing and movement have been adjusted, and elements now display the textual content more suitably. Additionally, a bug has been fixed where the translation of the value was incorrectly calculated.