Update package and enhance example interface
Updated package version with newly added description and fixed self-dependency issue. Also, the interface of the example was enhanced, with the adding of input fields for animation, random and reset options. Improvements were also made to the CSS file to style the new elements. A dist/examples directory was added to the .npmignore file and the baseUrl was changed in the tsconfig file.